Fix: Harmonize the meaning of LTTNG_EVENT_ALL
Analysis of the enable/disable code paths of the lttng-sessiond
indicates that LTTNG_EVENT_ALL is interpreted completely differently
depending on whether or not an event belongs to the UST, Python, JUL,
Log4j or kernel domain.

While all userspace domains interpret this to mean "any" type of
instrumentation, the kernel domain seems to interpret its meaning as
"enable all events".

In the spirit of presenting a sane-ish API which may be understood by
users who don't possess an intimate understanding of the session
daemon's intricacies, a careful statistical analysis was undertaken and
it was determined that since 4/5 domains are mutualy-consistent, the
remaining domain must be buggy.

This is marked as a "Fix" since this inconsistency made it impossible
to disable a specific kernel event.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>