Fix: hang of session daemon on kernel consumer launch
The kernel consumer does not handle command replies the same way
as the user space consumer. In the kernel consumer's case, a
reply must be sent explicitly by each command implementations
whereas a generic reply is sent everytime by the user space
consumer.

This caused the session daemon to block on a recvmsg() call
while it was waiting for the consumer to signal whether or not
the command (which was not implemented yet), was successful.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>