WIP fix: evaluate condition/trigger on subscription
If evaluation fails, for now, we follow through (RFC) with registration
(add to client list).

For now the evaluation requires a crude search to find the
channel associated with the trigger. A fast lookup would be
possible with the maintaining of a hash table with the relation trigger
-> channel. (RFC)

Evaluation is performed in handle_evaluation_on_subscription. Not sure
what would be a proper name here. "handle" seems to be reserved for
first level command response. (RFC)

Function prototype were added for convenience. Should this be done for
other static functions inside the file?

 - hash_channel_key
 - evaluate_condition
 - send_evaluation_to_clients

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>