Support for the metadata cache dump kernel command
This command allows the consumer daemon to reset the read position of
the metadata cache to 0. That way, it can be read again. This is used by
the session rotation feature to get a new copy of what was in the
metadata cache without regenerating it and re-sampling the offset from
epoch.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>