Cleanup subdir for CONSUMER_DST_NET
The subdir in the consumer_output structure is supposed to be the
domain-specific sub-directory (/kernel or /ust), but currently we
prepend <hostname>/<session-name> to this subdir for sessions that use a
relay.

Here, we introduce the dst.net.base_dir to store the
<hostname>/<session-name> in the consumer_output and keep subdir
only for the domain-specific data.

This is useful for session rotation where we need to insert the
chunk-name between the session-name and the domain-specific folder.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>