run_as: add return value abstraction
In the past, run_as commands were simple commands all returning int
values (e.g. open(), mkdir()). With the introduction of the future, more
complex run_as commands as the one used to extract elf symbol offsets
for the userspace-probe instrumentation we need to allow other return
 types such as long.

Creating an return type abstraction allows to add new return types. To
implement this new abstraction, this commit adds the passing of the
run_as_ret structure as parameter to the command functions that are
responsible to set the errno and the return value according to their
semantics. The run_as_ret struct now contains an union of return types.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>