Rotation on the relay
Support for the RELAYD_ROTATE_STREAM command on the relay. This command
informs the relay that the current stream must rotate after it has
written the data and index for the net_seq_num passed. After each data
and index written on disk we check if it is time to rotate (in case it
was in flight when the rotate command was received). On the other hand,
if too much data has been written when we receive the rotate command, we
move the excess data to a new tracefile and truncate the current one.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>