Rotate pending support on the relay
This command allows the sessiond to check if a rotation is complete from
the relayd point of view. There can be a significant delay between the
time the consumer has finished extracting the data from the buffers and
the time the relay has finished writing them on disk, and we can only
inform the user that the rotation is complete when all the data is on
disk. So the RELAYD_ROTATE_PENDING command is used to poll the relayd
after the consumer has finished extracting the data until everything is
on the relayd disk.

This command also takes care of streams that did not exist on the
consumer when the rotation started, or streams that appeared after the
last rotation started. The chunk_id field is used to distinguish those
cases.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>