Prevent manual session rotation when auto rotation is configured
This prevents a user from calling "lttng rotate" when the session has a
size or timer-based rotation configured. There is no clear use-case for
supporting that as the resulting chunk content would be random depending
on when the previous automatic rotation occured. Also, since it is
possible to disable automatic rotation settings at run-time, the user
can still trigger manually a rotation if needed.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>