Add the GMT offset in the rotated chunk path
The path of a rotated chunk is composed of the start and end timestamp
of the trace inside that chunk. In order to support distributed
environments, we now specify the GMT offset in this path as well. Here
is an example:
~/lttng-traces/<session-name>/20180111-134227-20180111-134228-GMT-5-1

Chunk start: 2018-01-11 13:42:27
Chunk end: 2018-01-11 13:42:28
GMT offset: GMT-5
Chunk ID: 1 (number of rotations that occured in this session so far)

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>