Notification: create and destroy session_info from command events
The notification thread maintains its hashtable of session_info
structures from the various channel creation and destruction events.

This was fine since all notifications so far only made sense as long
as a session had actual channels. The rotation notifications now
make it possible to get internal events related to a session that
has no channels. For instance, a rotation can be performed on a
session that has no channels active.

Example:

$ lttng create my_session
$ lttng enable-event -u -a
// No application running
$ lttng rotate

The notification thread now listens for the session creation
and destruction commands which allows us to handle such future
use-cases.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>