Print the git version used to build from a distribution tarball
The git version is omitted when building from a distribution
tarball. This will cause 'lttng version' and 'lttng --version'
to print the state of the git tree which produced the tarball.

git describe is used to produce the description of the tree's
state, along with the "dirty" state (whether or not local
changes were present in the tree).

Note that the 'git version' will not be printed when the
distribution tarball was produced at a release tag (a tag
starting with v[0-9]).

This patch simplifies the generation of the version.h file by
generating a file that is merely included by version.h.

It also ensures that version.tmpl is no longer installed on the
system by the install target.

Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>