Reference channels by their key in consumer timer management
The consumer's timer thread relies on SIGEVENT's sival_ptr value
to access a channel on which a periodic action applies.

This introduces a number of workarounds to ensure that the channel
is not accessed through a pending signal after its destruction.

In essence, all signals of a given type must be flushed before
the consumer can be certain that none are pending for a channel
it is destroying.

In scenarios where signal processing is slowed down, or if a
large number of channels are enabled, this can take a long time
or never complete as signals of a given type (live, switch, monitor)
are always pending.

This was observed on a slow network since the "live" signal
takes longer to process than its period. This resulted in a
live signal always being pending, preventing any progress.

Referencing channel through their key allows the consumer to
dispose of channels safely even if signals may be queued. The
consumer timer thread only relies on the RCU guarantees to
ensure the channel exists and will ignore the signals that
refer to non-existant channels.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>