Add the trace chunk and trace chunk registry interfaces
A trace chunk is a set of stream files. It maps to the user-visible
concept of "trace archive chunks" produced following a tracing session
rotation.

The concept of a "chunk" is introduced to make it possible/easy to
associate a group of stream files together, store common
properties (e.g. the epoch, base path, list of files), and perform an
action once all files have been closed/released.

The "trace chunk" interface is to be used by both the consumer and
relay daemons.

lttng_dynamic_pointer_array and lttng_dynamic_array utils are added
since they are used by the trace chunk implementation.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>